Add comprehensive tests for IngredienteViewSet and fix serializer in views
A variety of unit tests for the `IngredienteViewSet` class in `ingredienti/tests.py` have been added, including the test setups and specific tests for list, create, retrieve, update, partial update, delete actions and the testing of associated recipes to an ingredient. This ensures more robust testing for handling of ingredients in the application. Additionally, a minor fix has been applied in `ingredienti/views.py` by correctly using the `RicettaSerializer` in `ricette` method instead of `IngredienteSerializer`.